==> scripts/code/03b_CSI_online_typing_automatic_answer_classification.Rmd <==
---
title: '03b CSI online typing: Automatic answer classification'
author: "Kirsten Stark"
date: "`r format(Sys.time(), '%d %B, %Y')`"
output: github_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

## Load packages

```{r load_packages}
rm(list = ls())

library(tidyr)
library(dplyr)
library(stringr)
library(stringdist)

options( "encoding" = "UTF-8" )
```

## Load data

```{r load_data}
# input 
input = "data_long_final.csv"

# input synonym/alternative naming list
alternatives = "naming_alternatives.csv"

# load data
df <- read.csv(here::here("data", "transient_data_files", input))

# load alternatives
alternatives <- read.csv(here::here("data", "supplementary_info", alternatives), 
                         sep = ";")
```

## Define functions
### 1) Function to delete of last character typed words if those are space or enter
This function checks whether the last character(s) of the word entries is a space or is the typed word "Enter", and if so, deletes the last/the last five characters. If the last character(s) are neither of both, the word remains unchanged. The function can be used within dplyr's mutate function. 
Additionally, the function has the option to delete an alternative ending, while keeping " " and "Enter" at the end of the word

```{r function_delete_ending}
delete_ending <- function(word, ending = NA) {
  if(is.na(ending)) {
      # if no custom ending is given, delete " " and "Enter" at the end
      case_when(str_ends(word, " ") ~ str_sub(word, end=str_length(word)-str_length(" ")),
            str_ends(word, "Enter") ~ str_sub(word, end =str_length(word)-str_length("Enter")),
            TRUE ~ word)
  } else{
      # else delete custom ending
      case_when(str_ends(word, ending) ~ str_sub(word, end=str_length(word)-str_length(ending)),
            TRUE ~ word)
  }
}
```

### 2) Replace special keys (e.g. backspace, shift, etc.) by other characters (e.g. numbers)
Function takes as entries the word entries, the keys to be changed, and the characters they should be replaced with.

```{r function_replace_special_chars}
replace_special_chars <- function(input, oldnames, newnames) {
  if(length(oldnames) != length(newnames)){
    print("Your oldname/newname vectors don't have the same length. Please correct!")
    stop()
  }
  for (i in 1:length(input)) {
    for (j in 1:length(oldnames)) {
      # loop through all input values and all special characters and replaces all special characters by new name
      input[i] <- str_replace_all(input[i], pattern = oldnames[j], 
                        replacement = as.character(newnames[j]))
      if( i == 1) {
          print(paste0("The pattern ", oldnames[j], 
                   " has been replaced by the pattern ", 
                   newnames[j], ".", sep = ""))}
    }}
  return(input)
} 


```

### 3) Function that computes the final words by applying all backspaces
The function takes as input the word entries and, optionally, the backspace identifier.

```{r function_replace_backspace}
replace_backspace <- function(input, backspace = "Backspace") {
  for(i in 1:length(input)) {
    # loop through all word entries and count number of backspaces in the current word
    backspaces <- str_locate_all(input[i], backspace)[[1]]
    for(j in 1:nrow(backspaces)){
      # loop as many times as there are backspaces
      # for the current first backspace, delete backspace and the preceding character
      input[i] <- sub(str_c(".{1}",backspace), "", input[i])
    }
  }
  return(input)
}
```

### 4) Function that computes the fuzzy string matching
Calculate stringdistance between (backspace corrected) input word and item/alternative namings, and select the "best match", i.e. the item/alternative with the lowest distance and the first letter being correct.  
The default method is the Jaro distance (Jaro-Winkler distance ("jw") with p = 0), but other methods, of the stringdist function (van der Loo, 2014) are theoretically possible as well, but further options of the stringdist function might be necessary to adapt as well.

```{r function_calculate_stringdist}
# alternatives should be a dataframe with at least one column called item (same entry for as many alternatives as there are for a specific item and another column called alternatives). 

calculate_stringdist <- function(word, stims, alternatives =
                                alternatives, method = "jw", p = 0,
                                weight = c(1,1,1,1), q = 1,
                                firstlettercorrect = TRUE) {
 
  # input check
  if(length(word) != length(stims)){
    print("Your typed words and item vectors don't have the 
          same length. Please correct!")
    stop()
  }
  
  # Compute string distance between word entry and item 
  # using given method
  stringdistance <- stringdist(toupper(word), 
                               toupper(stims), method = method,
                               p = p, weight = weight, q = q)
  
  # placeholders that will be filled in the for-loop
  bestmatch <- stims
  dist <- 100 
  
  # convert all entries to upper
  word <- toupper(word)
  stims <- toupper(stims)
  alternatives$item <- toupper(alternatives$item)
  alternatives$alternatives <- toupper(alternatives$alternatives)
  
  # loop through all word entries
  for(i in 1:length(word)){
    # loop only if string distance to item is not already perfect
    # and is distance is not NA (meaning the word is NA)
    if(stringdistance[i] != 0 & !is.na(stringdistance[i])) {
      # curritem <- alternatives %>% filter(item == stims[i]) %>% 
      #   mutate(dist = stringdist(word[i], alternatives,
      #               method = method, p = p, q = q, 
      #               weight = weight)) %>%
      #   filter(substring(word[i],1,1) == 
      #        substring(alternatives,1,1)) %>%
      #   slice(which.min(dist)) 
      # if(nrow(curritem) != 0 &
      #    curritem$dist[1] < stringdistance[i]){
      #   stringdistance[i] <- curritem$dist[1]
      #   bestmatch[i] <- curritem$alternatives[1]
      # }}}

      # filter "alternatives" df for alternatives of current item
      curritem <- alternatives %>% filter(item == stims[i])
      # check whether current alternative column is not empty
      if(nrow(curritem) != 0) {
        for(j in 1:nrow(curritem)) {
          currentalternative <- curritem$alternatives[j]
          dist <- stringdist(word[i], currentalternative,
                    method = method, p = p, q = q, weight = weight)
          # compare the current string distance to the best
          # distance so far
                # for firstlettercorrect = TRUE
          if(dist < stringdistance[i] &
             firstlettercorrect == TRUE  &
             substring(word[i],1,1) ==
             substring(currentalternative,1,1)) {
                  stringdistance[i] <- dist
                  bestmatch[i] <- currentalternative
                  dist <- 100
                  } else if (dist < stringdistance[i]) {
                # for firstlettercorrect = FALSE
                  stringdistance[i] <- dist
                  bestmatch[i] <- currentalternative
                  dist <- 100
                  }
        }}}}
  distancebest <- cbind(stringdistance, bestmatch)
  return(distancebest)
}

```

### 5) Function that classifies the word entries
Function that classifies the word entries for correctness and different typing errors.

```{r classification_function}
case_character_type <- function(word, item, wordcorrected,
                                distance, bestmatch, d) {
  case_when(
        
    # correct answers: participants typed exactly the correct word, 
    # with space or enter at the end
     toupper(word) == toupper(item) | 
       toupper(word) == toupper(str_c(item, " ")) |
       toupper(word) == toupper(str_c(item, "Enter")) ~ "correct",
     
    # correctedtocorrect: participants corrected their entry to the correct
    # word using "Backspace"
    (toupper(wordcorrected) == toupper(item) | 
       toupper(wordcorrected) == toupper(str_c(item, " ")) |
       toupper(wordcorrected) == toupper(str_c(item, "Enter"))) &
        substring(wordcorrected,1,1) == substring(word,1,1) &
        substring(word,2, 10) != "Backspace" ~ "correctedtocorrect",
    
    # approx_correct: the approximately correct and best fitting word is the actual item
    # distance limits needs to be set
    (distance < d) & toupper(item) == toupper(bestmatch) & 
      toupper(substring(wordcorrected,1,1)) ==
                  toupper(substring(bestmatch,1,1)) & 
      toupper(substring(word, 1,1)) == 
              toupper(substring(bestmatch,1,1)) & 
      substring(word,2, 10) != "Backspace" ~ "approx_correct",
    
    # alternative: alternative was typed correctly 
      (distance == 0) & (toupper(word) == toupper(wordcorrected) |
        toupper(str_c(word, " ")) == toupper(wordcorrected) |
        toupper(str_c(word, "Enter")) == toupper(wordcorrected))  &
      toupper(substring(word,1,1)) ==
      toupper(substring(bestmatch,1,1)) ~ "alternative",
    
    # alternative_corrected: alternative typed correctly after backspace correction
     (distance == 0) & toupper(word) != toupper(wordcorrected) &
      toupper(str_c(word, " ")) != toupper(wordcorrected) &
      toupper(str_c(word, "Enter")) != toupper(wordcorrected) &
      toupper(substring(word, 1,1)) ==
      toupper(substring(bestmatch,1,1)) &
      substring(word,2, 10) != "Backspace" ~ 
      "alternative_corrected",
    
    # approx_alternative: distance limit needs to be set
    (distance < d) & distance != 0 &
      toupper(substring(wordcorrected,1,1)) ==
      toupper(substring(bestmatch,1,1)) &
      toupper(substring(word, 1,1)) ==
      toupper(substring(bestmatch,1,1)) &
      substring(word,2, 10) != "Backspace" ~ "approx_alternative",
       
    # backspace_space_enter: participants started by typing backspace, space,
    # enter, or capslock
     str_starts(word,"Backspace") |
       str_starts(word," ") |
       str_starts(word,"CapsLock") |
       str_starts(word,"Enter") ~ "backspace_space_enter",
    
    # shift_start: participants started by pressing the shift key
     str_starts(word,"Shift") ~ "shift_start",
    
    # isna: participants didn't enter anything
      is.na(word)       ~ "isna",
    
    # distance-based error
    (distance >= d) & 
      toupper(substring(wordcorrected,1,1)) ==
      toupper(substring(bestmatch,1,1)) &
      toupper(substring(word, 1,1)) ==
      toupper(substring(bestmatch,1,1)) &
      substring(word,2, 10) != "Backspace" ~ "distance_based_error",
    
    # first letter-based error
    (distance < d) & 
      (toupper(substring(wordcorrected,1,1)) !=
      toupper(substring(bestmatch,1,1)) |
      toupper(substring(word, 1,1)) !=
      toupper(substring(bestmatch,1,1)) |
      substring(word,2, 10) == "Backspace") ~ "first_letter_error",
  
    # are all answers classified
    TRUE                      ~ "not_correct" )
}
```


## Preprocess data, applying functions
### 1) Clean word ending 
By deleting the last character(s) of typed words if those are space or enter keys. (Alternatively, the function also takes custom endings that should be deleted.)  
As entries, the delete_ending function takes the column with the word entries and, optionally, a custom ending.
We can repeat applying this function if we want to keep deleting if Enter or space is repeated several times at the end of the word. The while loops stops as soon as none of the words has a space or Enter (or custom ending) at the end. (In our case, this changes only the ending of three words)

```{r clean_word_ending}
#df2 <- df %>% mutate(word.c = delete_ending(df$word))
isnotequal <- 1
df$word.c = currentupdate = df$word
while (isnotequal > 0) {
  df <- df %>% mutate(word.c = delete_ending(df$word.c))
  isnotequal <- sum(currentupdate != df$word.c, na.rm = TRUE)
  currentupdate <- df$word.c
}
# df2 <- df %>% mutate(word.c = delete_ending(df$word, ending = " "))
# df2 <- df %>% mutate(word.c = delete_ending(df$word.c, ending = "Enter"))
# sum(df$word.c != df2$word.c, na.rm = T)
# df2$word[df$word.endc != df2$word.c & !is.na(df$word)]
```

### 2) Replace special characters
Special characters such as Enter and Backspace are written as entire words. We want to replace these with identifiable numbers.

```{r replace_special_chars}
oldnames <- c("Enter", "CapsLock", "Shift", "ArrowLeft", "ArrowRight", "Backspace", "Control")
newnames <- c("1", "2", "3", "4", "5", "6", "7")
df$word.c <- replace_special_chars(input = df$word.c, oldnames = oldnames, newnames = newnames)
df$word.cc <- df$word.c
```

### 3) Compute finally submitted words by applying all backspaces
Function takes as input the word entries and, optionally, the backspace identifier. 

```{r replace_backspace}
#df$word.c[1:200]
#df$word.cb <- replace_backspace(df$word.c, backspace = "Backspace")
df$word.c <- replace_backspace(df$word.c, backspace = "6")
#df$word.cb[1:200]
```

### 4) Compute stringdist between word entries and items/alternatives
Compute Jaro distance 

```{r calculate_stringdist_jaro}
#stringdist(toupper(df$word.c2[1:200]), toupper(df$item[1:200]), method = "jw")
tictoc::tic()
output <- calculate_stringdist(word = df$word.c, stims = df$item, 
                               alternatives = alternatives, 
                               method = "jw", p = 0,
                               firstlettercorrect = TRUE)
tictoc::toc()
df$jaro <- output[,1]
df$bestmatch_jaro <- output[,2]
#df$jaro[1:200]
```

### 5) Classify word entries

```{r classify_entries}
df2 <- df %>% 
  mutate(answer_auto_jaro = case_character_type(word, item, 
          word.c, jaro, bestmatch_jaro, d = 0.3))
```

```{r classify_as_correct}
df2 <- df2 %>% 
  mutate(correct_auto_jaro = case_when(
    answer_auto_jaro == "correct" ~ 1,
    answer_auto_jaro == "correctedtocorrect" ~ 1,
    answer_auto_jaro == "approx_correct" ~ 1,
    answer_auto_jaro == "alternative" ~ 1,
    answer_auto_jaro == "alternative_corrected" ~ 1,
    answer_auto_jaro == "approx_alternative" ~ 1,
    TRUE ~ 0)) %>%
  mutate(correct_manual = case_when(correct == 1 ~ 1,
                                    is.na(correct) ~ 0))
```


### Inspect results

```{r}
new_correct <- df2 %>% 
  filter(correct_manual == 0 &
           (correct_auto_jaro == 1)) %>%
  select(item, word, word.c, bestmatch_jaro, answercode, answer_auto_jaro)

new_incorrect <- df2 %>% 
  filter(correct_manual == 1 &
           (correct_auto_jaro == 0)) %>%
  select(item, word, word.c, bestmatch_jaro, answercode, answer_auto_jaro)

both_incorrect <- df2 %>% 
  filter(correct_manual == 1 & correct_auto_jaro == 1) %>%
  select(item, word, word.c,bestmatch_jaro, jaro, answercode,answer_auto_jaro)
                                  
print("Jaro vs. manual: ");
table(df2$correct_auto_jaro, df2$correct_manual);
```

## Comparison manual and automatic classification (Jaro distance)

```{r inspection_manual_vs_automatic}
table(df2$correct_auto_jaro, df2$correct_manual)
print("The new correct trials:")
(new_correct <- df2 %>% filter(correct_manual == 0 & correct_auto_jaro == 1) %>%
  select(item, word, word.c, bestmatch_jaro, jaro,
         answer_auto_jaro, answercode))
print("Amount of trials additionally considered as correct: ");
sum(df2$correct_manual == 0 & df2$correct_auto_jaro == 1);
print("In percent: ");
round(sum(df2$correct_manual == 0 & df2$correct_auto_jaro == 1)/
        nrow(df2)*100,2)

print("The new incorrect trials:")
(new_incorrect <- df2 %>% filter(correct_manual == 1 & correct_auto_jaro == 0) %>%
  select(item, word, word.c, bestmatch_jaro, jaro,
         answer_auto_jaro, answercode))
print("Amount of trials additionally considered as incorrect: ");
sum(df2$correct_manual == 1 & df2$correct_auto_jaro == 0);
print("In percent: ");
round(sum(df2$correct_manual == 1 & df2$correct_auto_jaro == 0)/
        nrow(df2)*100,2)
```

**Total amount of trials classified differently (in percent):**

```{r total_difference}
round(((sum(df2$correct_manual == 0 & df2$correct_auto_jaro == 1)+ sum(df2$correct_manual == 1 & df2$correct_auto_jaro == 0))/
  nrow(df2))*100,2)
```

**Correlation between manual and automatic classification:**

```{r correlation_manual_automatic}
round(cor(df2$correct_manual, df2$correct_auto_jaro),2)
```

## Typing error analyses based on automatic classification

Amount of trials classified as correct and incorrect
```{r table_correct_incorrect_jaro}
print("totaltrials:")
nrow(df2)
print("correct:")
(correct = sum(df2$correct_auto_jaro == 1))
print("incorrect:")
(incorrect = sum(df2$correct_auto_jaro == 0))
```

Percentage of incorrect trials

```{r percentage_incorrect}
# incorrect/nrow(df2)*100
# incorrect/30/160*100
incorrect_per_subject <- 
  as.data.frame(table(df2$subject, df2$correct_auto_jaro)) %>%
  filter(Var2 == 0) %>% select(Var1, Freq) %>%
  dplyr::rename(subject = Var1, perct_incorrect = Freq) %>%
  mutate(perct_incorrect = perct_incorrect/160)

print("Mean:")
round(mean(incorrect_per_subject$perct_incorrect)*100,2)
print("SD:")
round(sd(incorrect_per_subject$perct_incorrect)*100,2)
print("Range:")
round(range(incorrect_per_subject$perct_incorrect)*100,2)


```

Correct/incorrect trials per participant: 

```{r}
print(as.data.frame(table(
  df2$subject, df2$correct_auto_jaro == 1)) %>%
    filter(Var2 == TRUE) %>%
    dplyr::rename(subject = Var1, totaltrials = Var2, 
                  correct_auto = Freq) %>%
    mutate(totaltrials = 160) %>% 
    mutate(percentagecorrect = correct_auto/totaltrials))
```

## Write data file for statistical analyses
```{r}
write.csv(df2, here::here("data","transient_data_files", "data_long_final.csv"))
```
