==> scripts/code/06_CSI_online_typing_comparison_to_verbal_naming.Rmd <==
---
title: "06 CSI online typing: Comparison to verbal naming"
author: "Kirsten Stark"
date: "`r format(Sys.time(), '%d %B, %Y')`"
output: github_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

# Load packages

```{r load packages}
library(dplyr)
library(tidyr)
library(lme4)
library(lmerTest)
library(Rmisc)
library(Cairo)
#library(strengejacke)
library(ggplot2)
library(sjPlot)

options(scipen=999)

rm(list = ls())
options( "encoding" = "UTF-8" )
set.seed(99)
```

# Load data
Load data from both the verbal online CSI experiment and the typing online CSI experiment.  
  
Load typing data

```{r load_typing_data}
df_typing <- read.csv(here::here("data", "data_long_anonymous.csv"))
```

Select classification type (as in script 05)

```{r classification_type}
classification_type <- "automatic"
if(classification_type == "automatic") {
  df_typing$answercode <- df_typing$answer_auto_jaro
  df_typing$correct <- df_typing$correct_auto_jaro
} else if(classification_type == "manual") {
  df_typing$answercode <- df_typing$answercode
  df_typing$correct <- df_typing$correct_manual
} else {
  print("Select a correct type!")
}
```

Load verbal data

```{r load_verbal_data}
load(here::here("data", "verbal_CSI", "CSI_online_verbal_df_full.RData"))
df_verbal <- df_full
```

# Combine both data frames into one
1) Subset relevant columns and give identical names

```{r subset_and_rename_columns}
df_typing <- df_typing %>% 
  dplyr::select(subject, item, category, timing.01, PosOr, correct) %>%
  dplyr::rename(RT = timing.01, Pos = PosOr) %>%
  mutate(experiment = "typing")

df_verbal <- df_verbal %>%
  dplyr::select(VP, Item, subcat, VOT, correct, Pos) %>%
  dplyr::rename(subject = VP, item = Item, category = subcat, RT = VOT) %>%
  mutate(experiment = "verbal")
```

2) Give subjects from both experiments different names

```{r adapt_subject_names}
df_typing <- df_typing %>% mutate(subject = subject + 200)
df_verbal <- df_verbal %>% mutate(subject = subject + 100)
```

3) Put columns into correct format 

```{r factorize}
df_typing <- df_typing %>% 
  mutate(subject = as.factor(subject)) %>%
  mutate(item = as.character(item)) %>%
  mutate(category = as.factor(category)) %>% 
  mutate(RT = as.numeric(RT)) %>%
  mutate(Pos = as.numeric(Pos)) %>% 
  mutate(experiment = factor(experiment, levels = c("verbal", "typing")))

df_verbal <- df_verbal %>% 
  mutate(subject = as.factor(subject)) %>%
  mutate(item = as.character(item)) %>%
  mutate(category = as.factor(category)) %>% 
  mutate(RT = as.numeric(RT)) %>%
  mutate(Pos = as.numeric(Pos)) %>% 
  mutate(experiment = factor(experiment, levels = c("verbal", "typing")))
```

4) Bind both data frames into one

```{r combine_df}
df <- bind_rows(df_typing, df_verbal)
```

5) Give identical category names in both experiments

```{r category_names}
df <- df %>% dplyr::mutate(category = case_when(category == "Buero" ~ "Büro",
                                  category == "Gebaeude" ~ "Gebäude",
                                  category == "Gemuese" ~ "Gemüse",
                                  category == 
                                    "Koerperteile" ~ "Körperteile",
                                  category == "Kueche" ~ "Küche",
                                  category == 
                                    "Suessigkeiten" ~ "Süssigkeiten",
                                  category == 
                                    "Trinkgefaesse" ~ "Trinkgefässe",
                                  category == "Voegel" ~ "Vögel",
                                  TRUE ~ as.character(category))) %>%
  mutate(category == as.factor(category)) %>% droplevels()
table(df$category)
```

5) Drop filler trials

```{r drop_filler}
 df <- df %>% filter(category != "Filler" & 
                      category != "Filler1" & category != "Filler2") %>%
  droplevels()
```

6) Drop incorrect trials

```{r subset_to_correct}
df <- df %>% filter(!is.na(correct) & correct != 0) %>% 
  dplyr::select(-correct) %>%
  droplevels()
```

7) Export combined data frame for post-hoc power plot

```{r export_df}
write.csv(df, here::here("data", "CSI_online_combined.csv"))
```

# Descriptives

```{r descriptives}
(descriptives <- df %>% 
   Rmisc::summarySEwithin(.,"RT",idvar = "subject",
                          withinvars = "Pos",
                          betweenvars = "experiment",
                          na.rm = T))
```

# Plotting
Make plots suitable for APA format

```{r apatheme}
apatheme <- theme_bw()+
  theme(plot.title=element_text(family="Arial",size=22,hjust = .5),
        panel.grid.major=element_blank(), panel.grid.minor=element_blank(),
        panel.border=element_blank(),axis.line=element_line(),
        text=element_text(family="Arial",size=16))
```

Plot RTs by ordinal position for both experiments

```{r plot}
(plot <- df %>% 
    ggplot(., aes(x=Pos, y=RT, 
                  group=experiment, color=experiment)) +
    stat_summary(fun=mean,  geom="point", size = 2)+
    stat_summary(fun=mean,  geom="line", size = 1) +
    apatheme+
    labs(x="Ordinal Position ",y ="RT (ms)", color = "Experiment"))

filename <- "CSI_online_exp_comparison_plot_rt.pdf"
ggsave(plot, filename = 
         here::here("results", "figures", filename),
       width = 18, height = 12, units = "cm", 
       dpi = 300, device = cairo_pdf)
embedFonts(file = here::here("results", "figures", filename))
```

# Set contrasts and convert continuous predictor variable

1) Set contrasts: Sliding difference contrast for factor experiment  
Subsequent factor levels are being compared to each other, i.e. verbal-typing, the intercept being the grand mean.

```{r set contrasts}
# define sliding difference contrast for factor experiment: 
# contrast is verbal - typing, intercept being the grand mean 
levels(df$experiment)
contrasts(df$experiment) <- MASS::contr.sdif(2)
```

2) Center the continuous predictor variable position around zero (ordinal contrast with five levels)

```{r Pos_cont}
df$Pos.cont <- scale(as.numeric(as.character(df$Pos)),
                       center = T, scale = F)
# table(df$Pos.cont)
# mean(df$Pos.cont)
```

# Check distribution of data
Are the data normally distributed or does a gamma distribution fit the data better?  
*Histogram of the reaction time data*

```{r RT_hist}
hist(df$RT)
```

*Check fit of normal vs gamma distribution in histograms, q-q-plots and using objective criteria:*  
1) Fit normal and gamma distributions to the reaction time data

```{r load_fitdistrplus}
library(fitdistrplus)
```

```{r fit.normal}
fit.normal<- fitdist(df$RT, distr = "norm", method = "mle")
summary(fit.normal)
#plot(fit.normal)
```

```{r fit.gamma}
fit.gamma <- fitdist(df$RT, distr = "gamma", method = "mle")
summary(fit.gamma)
#plot(fit.gamma)
```

2) Compare the fit of the two distributions  
Visually compare fit of both distributions in histogram

```{r density-comparison}
denscomp(list(fit.gamma, fit.normal))
```
Visually compare fit of both distributions in Q-Q-plots

```{r q-q-comparison}
qqcomp(list(fit.gamma, fit.normal))
```

Compare information criteria

```{r fit_criteria-comparison}
gofstat(list(fit.gamma, fit.normal),
        fitnames = c("Gamma", "Normal"))
```

**Conclusion:** Both the visual inspection and the objective criteria suggest that a gamma distribution fits the data better. Therefore, we fit a Gamma distribution in a GLMM with the continuous predictor ordinal position (Pos.cont), the factorial predictor (experiment), and their interaction. We compute the maximal random effects structure.

# Interferential statistics: GLMM with predictors Pos.cont and experiment

```{r glmm_cont}
m1 <- afex::lmer_alt(RT ~ Pos.cont*experiment +
               (Pos.cont||subject) +(Pos.cont*experiment||category),
             data = df,
            family =Gamma(link ="identity"),
            control=glmerControl(optimizer = "bobyqa", optCtrl = list(maxfun = 2e5)))
summary(m1)

# save model output
tab_model(m1,transform = NULL,
          show.re.var = F, show.stat = T,show.r2 = F,show.icc = F,
          title = "GLMM (Gamma distribution)",
          pred.labels = c("(Intercept)", "Ordinal Position", 
                          "Experiment", "(Ordinal Position) x 
                          \n (Experiment)"),
          wrap.labels = 10,
          dv.labels = "Typing Onset Latency",
          #string.pred = "",
          string.stat = "t-Value",
          file = here::here("results", "tables",
                            "CSI_online_experiment_comparison_glmm_cont.html"))
```
