==> scripts/code/07a_CSI_online_post-hoc_power_simulations.Rmd <==
---
title: "07a CSI online additional analyses: Post-hoc power"
author: "Kirsten Stark"
date: "`r format(Sys.time(), '%d %B, %Y')`"
output: github_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

## Load packages

```{r load_packages,results="hide", cache = TRUE}
library(dplyr)
library(tidyr)
library(devtools)
library(MASS)
library(lme4)
library(lmerTest)
library(simr)
library(pbkrtest)
library(testthat)
library(ggplot2)

rm(list = ls())

set.seed(99)
```

## Load data

```{r load data}
input = "CSI_online_combined.csv"
df <- read.csv(here::here("data", input))

# factorize columns
is.numeric(df$RT)
df$subject <- as.factor(df$subject)
df$category <- as.factor(df$category)
df$experiment <- as.factor(df$experiment)
```

## 1) Set up models - LMM with log-transformed reaction times
Also the GLMM fit the data better, it doesn't work well with the simr package. 
Therefore, we fit the power analyses with LMMs instead. 
LMM with continuous predictor ordinal position and log-transformed RT as a
dependent variable.

```{r model_CSI_verbal}
# subset verbal data
df_verbal <- df %>% filter(experiment == "verbal") %>% droplevels()

# center continuous predictor 
df_verbal$Pos.cont <- scale(as.numeric(as.character(df_verbal$Pos)), 
                            center = TRUE, scale = FALSE)

# use log-transformed data because simr can't handle GLMM well
df_verbal$RT.log <- log(df_verbal$RT)

# fit model 
m_verbal <- lmer(RT.log ~ Pos.cont +
               (Pos.cont|subject) +(Pos.cont|category),
             data = df_verbal,
            control=lmerControl(optimizer = "bobyqa"))
summary(m_verbal)
```

```{r model_CSI_typing}
# subset verbal data
df_typing <- df %>% filter(experiment == "typing") %>% droplevels()

# center continuous predictor 
df_typing$Pos.cont <- scale(as.numeric(as.character(df_typing$Pos)), 
                            center = TRUE, scale = FALSE)

# use log-transformed data because simr can't handle GLMM well
df_typing$RT.log <- log(df_typing$RT)

# fit model 
m_typing <- lmer(RT.log ~ Pos.cont +
               (Pos.cont|subject) +(1|category),
             data = df_typing,
            control=lmerControl(optimizer = "bobyqa",
                                 optCtrl = list(maxfun = 2*10^5)))
summary(m_typing)
```

##2) Extend dataset
We do not need to extend for the post-hoc power analysis, and keep the 30 participants.

```{r extend_verbal_data}
m_verbal  <- extend(m_verbal, along="subject", n=30)
#m2_verbal_data <- getData(m2_verbal) 
### ok, data were indeed extended to n subjects ;-)
#str(m2_verbal_data)
#str(df_verbal)
```

```{r extend_typing_data}
m_typing  <- extend(m_typing, along="subject", n=30)
#m2_typing_data <- getData(m2_typing) 
### ok, data were indeed extended to n subjects ;-)
#str(m2_typing_data)
#str(df_typing)
```

## 3) Specify effect sizes
As we want to evaluate the post-hoc power for different sample sizes, we do not 
change the effect sizes

```{r}
fixef(m_verbal)
fixef(m_verbal)["Pos.cont"] <- 0.02756341  
fixef(m_verbal)["(Intercept)"] <- 6.84035775  
fixef(m_typing)
fixef(m_typing)["Pos.cont"] <- 0.02961954 
fixef(m_typing)["(Intercept)"] <- 7.07567744

```

## 4) Run the power analysis

We want to run power analyses at different sample sizes and with different trial numbers.

#### A) **VERBAL DATA**

*30 participants, 120 trials - The actual post-hoc power :*  

```{r PowerSim_m2_verbal_30,results="hide", cache = TRUE}
(Power_m2_verbal_30 <- powerSim(m_verbal, 
                                test=fixed("Pos.cont", method = "t"),
                                nsim = 1000)) # increase to nsim = 1000 for real test

saveRDS(Power_m2_verbal_30, 
        file = here::here("results", "power_sim",
                          "pc_verbal30.RDS"))

```

```{r Power_m2_verbal_30_save_for_export}
results <- summary(Power_m2_verbal_30)
Power_m2_verbal_30$warnings
Power_m2_verbal_30$errors

simr_results <- as.data.frame(cbind(rep("verbal_s30_t_around109", 
                                        times = Power_m2_verbal_30$n),
                                    rep(30, times =
                                          Power_m2_verbal_30$n),
                                    rep(109, times =
                                          Power_m2_verbal_30$n),
                                    rep(24, times = Power_m2_verbal_30$n),
              Power_m2_verbal_30$pval, results$successes,
              results$mean, results$lower, results$upper,
              rep(nrow(Power_m2_verbal_30$warnings), 
                  times =Power_m2_verbal_30$n)))
colnames(simr_results)<-c("name", "nsubjects", "ntrials", "ncat",
                          "pval",
                          "successes", "mean", "lower", "upper", "warnings")
``` 


Save data frame

```{r PowerSim_m2_verbal_30_export_data}
write.csv(simr_results, here::here("results", "power_sim",  "posthoc_powersim.csv"))
```


*24 participants, 120 trials:*


```{r PowerCurve_m2_verbal_24,results="hide", cache = TRUE}
(pc_verbal24 <- powerCurve(m_verbal, test=fixed("Pos.cont", method = "t"),
                           along = "subject", breaks = c(24), nsim = 1000))

saveRDS(pc_verbal24, 
        file = here::here("results", "power_sim",
                          "pc_verbal24.RDS"))
```

```{r PowerCurve_m2_verbal_24_results}
#plot(pc_verbal24)
summary(pc_verbal24)
pc_verbal24$warnings
pc_verbal24$errors
```

```{r PowerCurve_m2_verbal_24_add_to_df}
results <- summary(pc_verbal24)
results_df <- as.data.frame(cbind(rep("verbal_s24_t_around109", 
                                        times= results$trials), 
                                    rep(24, times = results$trials),
                                    rep(87, times = results$trials),
                                  rep(24, times = results$trials),
              pc_verbal24[["ps"]][[1]]$pval, 
              rep(results$successes, times = results$trials),
              rep(results$mean,  times = results$trials),
              rep(results$lower,  times = results$trials),
              rep(results$upper,  times = results$trials),
              rep(nrow(pc_verbal24$warnings), 
                  times = results$trials)))
colnames(results_df)<-c("name", "nsubjects", "ntrials", "ncat",
                        "pval","successes", "mean", "lower", "upper", "warnings")
``` 

```{r PowerCurve_m2_verbal_24_export}
simr_results <- rbind(simr_results, results_df)
write.csv(simr_results, here::here("results", "power_sim",  "posthoc_powersim.csv"))
```


*20 participants, 120 trials:*


```{r PowerCurve_m2_verbal_20,results="hide", cache = TRUE}
(pc_verbal20 <- powerCurve(m_verbal, test=fixed("Pos.cont", method = "t"),
                           along = "subject", breaks = c(20), nsim = 1000))

saveRDS(pc_verbal20, 
        file = here::here("results", "power_sim",
                          "pc_verbal20.RDS"))
```

```{r PowerCurve_m2_verbal_20_results}
#plot(pc_verbal20)
summary(pc_verbal20)
pc_verbal20$warnings
pc_verbal20$errors
```

```{r PowerCurve_m2_verbal_20_add_to_df}
results <- summary(pc_verbal20)
results_df <- as.data.frame(cbind(rep("verbal_s20_t_around109", 
                                        times= results$trials), 
                                    rep(20, times = results$trials),
                                    rep(73, times = results$trials),
                                  rep(24, times = results$trials),
              pc_verbal20[["ps"]][[1]]$pval, 
              rep(results$successes, times = results$trials),
              rep(results$mean,  times = results$trials),
              rep(results$lower,  times = results$trials),
              rep(results$upper,  times = results$trials),
              rep(nrow(pc_verbal20$warnings), 
                  times = results$trials)))
colnames(results_df)<-c("name", "nsubjects", "ntrials", "ncat",
                        "pval","successes", "mean", "lower", "upper", "warnings")
``` 

```{r PowerCurve_m2_verbal_20_export}
simr_results <- rbind(simr_results, results_df)
write.csv(simr_results, here::here("results", "power_sim",  "posthoc_powersim.csv"))
```


*16 participants, 120 trials:*

```{r PowerCurve_m2_verbal_16,results="hide", cache = TRUE}
(pc_verbal16 <- powerCurve(m_verbal, test=fixed("Pos.cont", method = "t"),
                           along = "subject", breaks = c(16), nsim = 1000))

saveRDS(pc_verbal16, 
        file = here::here("results", "power_sim",
                          "pc_verbal16.RDS"))
```

```{r PowerCurve_m2_verbal_16_results}
#plot(pc_verbal16)
summary(pc_verbal16)
pc_verbal16$warnings
pc_verbal16$errors
```

```{r PowerCurve_m2_verbal_16_add_to_df}
results <- summary(pc_verbal16)
results_df <- as.data.frame(cbind(rep("verbal_s16_t_around109", 
                                        times= results$trials), 
                                    rep(16, times = results$trials),
                                    rep(109, times = results$trials),
                                rep(24, times = results$trials),
              pc_verbal16[["ps"]][[1]]$pval, 
              rep(results$successes, times = results$trials),
              rep(results$mean,  times = results$trials),
              rep(results$lower,  times = results$trials),
              rep(results$upper,  times = results$trials),
              rep(nrow(pc_verbal16$warnings), 
                  times = results$trials)))
colnames(results_df)<-c("name", "nsubjects", "ntrials", "ncat",
                        "pval","successes", "mean", "lower", "upper",
                        "warnings")
``` 

```{r PowerCurve_m2_verbal_16_export}
simr_results <- rbind(simr_results, results_df)
write.csv(simr_results, here::here("results", "power_sim",  "posthoc_powersim.csv"))
```


*10 participants, 120 trials:*


```{r PowerCurve_m2_verbal_10,results="hide", cache = TRUE}
(pc_verbal10 <- powerCurve(m_verbal, test=fixed("Pos.cont", method = "t"),
                           along = "subject", breaks = c(10), nsim = 1000))

saveRDS(pc_verbal10, 
        file = here::here("results", "power_sim",
                          "pc_verbal10.RDS"))
```

```{r PowerCurve_m2_verbal_10_results}
#plot(pc_verbal10)
summary(pc_verbal10)
pc_verbal10$warnings
pc_verbal10$errors
```

```{r PowerCurve_m2_verbal_10_add_to_df}
results <- summary(pc_verbal10)
results_df <- as.data.frame(cbind(rep("verbal_s10_t_around109", 
                                        times= results$trials), 
                                    rep(10, times = results$trials),
                                    rep(36, times = results$trials),
                                  rep(24, times = results$trials),
              pc_verbal10[["ps"]][[1]]$pval, 
              rep(results$successes, times = results$trials),
              rep(results$mean,  times = results$trials),
              rep(results$lower,  times = results$trials),
              rep(results$upper,  times = results$trials),
              rep(nrow(pc_verbal10$warnings), 
                  times = results$trials)))
colnames(results_df)<-c("name", "nsubjects", "ntrials", "ncat",
                        "pval","successes", "mean", "lower", "upper", "warnings")
``` 

```{r PowerCurve_m2_verbal_10_export}
simr_results <- rbind(simr_results, results_df)
write.csv(simr_results, here::here("results", "power_sim",  "posthoc_powersim.csv"))
```


*8 participants, 120 trials:*

```{r PowerCurve_m2_verbal_8,results="hide", cache = TRUE}
(pc_verbal8 <- powerCurve(m_verbal, test=fixed("Pos.cont", method = "t"),
                           along = "subject", breaks = c(8), nsim = 1000))

saveRDS(pc_verbal8, 
        file = here::here("results", "power_sim",
                          "pc_verbal8.RDS"))
```

```{r PowerCurve_m2_verbal_8_results}
#plot(pc_verbal8)
summary(pc_verbal8)
pc_verbal8$warnings
pc_verbal8$errors
```

```{r PowerCurve_m2_verbal_8_add_to_df}
results <- summary(pc_verbal8)
results_df <- as.data.frame(cbind(rep("verbal_s8_t_around109", 
                                        times= results$trials), 
                                    rep(8, times = results$trials),
                                    rep(30, times = results$trials),
                                  rep(24, times = results$trials),
              pc_verbal8[["ps"]][[1]]$pval, 
              rep(results$successes, times = results$trials),
              rep(results$mean,  times = results$trials),
              rep(results$lower,  times = results$trials),
              rep(results$upper,  times = results$trials),
              rep(nrow(pc_verbal8$warnings), 
                  times = results$trials)))
colnames(results_df)<-c("name", "nsubjects", "ntrials", "ncat", 
                        "pval","successes", "mean", "lower", "upper", "warnings")
``` 

```{r PowerCurve_m2_verbal_8_export}
simr_results <- rbind(simr_results, results_df)
write.csv(simr_results, here::here("results", "power_sim",  "posthoc_powersim.csv"))
```

*6 participants, 120 trials:*


```{r PowerCurve_m2_verbal_6,results="hide", cache = TRUE}
(pc_verbal6 <- powerCurve(m_verbal, test=fixed("Pos.cont", method = "t"),
                           along = "subject", breaks = c(6), nsim = 1000))

saveRDS(pc_verbal6, 
        file = here::here("results", "power_sim",
                          "pc_verbal6.RDS"))
```

```{r PowerCurve_m2_verbal_6_results}
#plot(pc_verbal5)
summary(pc_verbal6)
pc_verbal6$warnings
pc_verbal6$errors
```

```{r PowerCurve_m2_verbal_6_add_to_df}
results <- summary(pc_verbal6)
results_df <- as.data.frame(cbind(rep("verbal_s6_t_around109", 
                                        times= results$trials), 
                                    rep(6, times = results$trials),
                                    rep(72, times = results$trials),
                                  rep(24, times = results$trials),
              pc_verbal6[["ps"]][[1]]$pval, 
              rep(results$successes, times = results$trials),
              rep(results$mean,  times = results$trials),
              rep(results$lower,  times = results$trials),
              rep(results$upper,  times = results$trials),
              rep(nrow(pc_verbal6$warnings), 
                  times = results$trials)))
colnames(results_df)<-c("name", "nsubjects", "ntrials", "ncat",
                        "pval","successes", "mean", "lower", "upper", "warnings")
``` 

```{r PowerCurve_m2_verbal_6_export}
simr_results <- rbind(simr_results, results_df)
write.csv(simr_results, here::here("results", "power_sim",  "posthoc_powersim.csv"))
```


*30 participants, 19 categories:*

```{r PowerCurve_m2_verbal_cat19,results="hide", cache = TRUE}
(pc_verbal_categories_19 <- powerCurve(m_verbal, test=fixed("Pos.cont", method = "t"),
                                       along = "category", breaks = c(19),
                                       nsim = 1000, seed = 99))

saveRDS(pc_verbal_categories_19, 
        file = here::here("results", "power_sim",
                          "pc_verbal_categories_19.RDS"))
```

```{r PowerCurve_m2_verbal_cat19_summary}
#plot(pc_verbal_categories_19)
summary(pc_verbal_categories_19)
pc_verbal_categories_19$warnings
pc_verbal_categories_19$errors
```

```{r PowerCurve_m2_verbal_cat19_add_to_df}
results <- summary(pc_verbal_categories_19)
results_df <- as.data.frame(cbind(rep("verbal_s30_cat19", 
                                        times= results$trials), 
                                    rep(30, times = results$trials),
                                    rep(87, times = results$trials),
                                    rep(19, times = results$trials),
              pc_verbal_categories_19[["ps"]][[1]]$pval, 
              rep(results$successes, times = results$trials),
              rep(results$mean,  times = results$trials),
              rep(results$lower,  times = results$trials),
              rep(results$upper,  times = results$trials),
              rep(nrow(pc_verbal_categories_19$warnings), 
                  times = results$trials)))
colnames(results_df)<-c("name", "nsubjects", "ntrials", "ncat",
                        "pval","successes", "mean", "lower", "upper", "warnings")
``` 

```{r PowerCurve_m2_verbal_cat19_export}
simr_results <- rbind(simr_results, results_df)
write.csv(simr_results, here::here("results", "power_sim",  "posthoc_powersim.csv"))
```


*30 participants, 16 categories:*

```{r PowerCurve_m2_verbal_cat16,results="hide", cache = TRUE}
(pc_verbal_categories_16 <- powerCurve(m_verbal, test=fixed("Pos.cont", method = "t"),
                                       along = "category", breaks = c(16),
                                       nsim = 1000, seed = 99))

saveRDS(pc_verbal_categories_16, 
        file = here::here("results", "power_sim",
                          "pc_verbal_categories_16.RDS"))
```

```{r PowerCurve_m2_verbal_cat16_summary}
#plot(pc_verbal_categories_16)
summary(pc_verbal_categories_16)
pc_verbal_categories_16$warnings
pc_verbal_categories_16$errors
```

```{r PowerCurve_m2_verbal_cat16_add_to_df}
results <- summary(pc_verbal_categories_16)
results_df <- as.data.frame(cbind(rep("verbal_s30_cat16", 
                                        times= results$trials), 
                                    rep(30, times = results$trials),
                                    rep(73, times = results$trials),
                                    rep(16, times = results$trials),
              pc_verbal_categories_16[["ps"]][[1]]$pval, 
              rep(results$successes, times = results$trials),
              rep(results$mean,  times = results$trials),
              rep(results$lower,  times = results$trials),
              rep(results$upper,  times = results$trials),
              rep(nrow(pc_verbal_categories_16$warnings), 
                  times = results$trials)))
colnames(results_df)<-c("name", "nsubjects", "ntrials", "ncat",
                        "pval","successes", "mean", "lower", "upper", "warnings")
``` 

```{r PowerCurve_m2_verbal_cat16_export}
simr_results <- rbind(simr_results, results_df)
write.csv(simr_results, here::here("results", "power_sim",  "posthoc_powersim.csv"))
```

*30 participants, 13 categories:*

```{r PowerCurve_m2_verbal_cat13,results="hide", cache = TRUE}
(pc_verbal_categories_13 <- powerCurve(m_verbal, test=fixed("Pos.cont", method = "t"),
                                       along = "category", breaks = c(13),
                                       nsim = 1000, seed = 99))

saveRDS(pc_verbal_categories_13, 
        file = here::here("results", "power_sim",
                          "pc_verbal_categories_13.RDS"))
```

```{r PowerCurve_m2_verbal_cat13_summary}
#plot(pc_verbal_categories_13)
summary(pc_verbal_categories_13)
pc_verbal_categories_13$warnings
pc_verbal_categories_13$errors
```

```{r PowerCurve_m2_verbal_cat13_add_to_df}
results <- summary(pc_verbal_categories_13)
results_df <- as.data.frame(cbind(rep("verbal_s30_cat13", 
                                        times= results$trials), 
                                    rep(30, times = results$trials),
                                    rep(59, times = results$trials),
                                    rep(13, times = results$trials),
              pc_verbal_categories_13[["ps"]][[1]]$pval, 
              rep(results$successes, times = results$trials),
              rep(results$mean,  times = results$trials),
              rep(results$lower,  times = results$trials),
              rep(results$upper,  times = results$trials),
              rep(nrow(pc_verbal_categories_13$warnings), 
                  times = results$trials)))
colnames(results_df)<-c("name", "nsubjects", "ntrials", "ncat",
                        "pval","successes", "mean", "lower", "upper", "warnings")
``` 

```{r PowerCurve_m2_verbal_cat13_export}
simr_results <- rbind(simr_results, results_df)
write.csv(simr_results, here::here("results", "power_sim",  "posthoc_powersim.csv"))
```


*30 participants, 8 categories:*

```{r PowerCurve_m2_verbal_cat8,results="hide", cache = TRUE}
(pc_verbal_categories_8 <- powerCurve(m_verbal, test=fixed("Pos.cont", method = "t"),
                                       along = "category", breaks = c(8),
                                       nsim = 1000, seed = 99))

saveRDS(pc_verbal_categories_8, 
        file = here::here("results", "power_sim",
                          "pc_verbal_categories_8.RDS"))
```

```{r PowerCurve_m2_verbal_cat8_summary}
#plot(pc_verbal_categories_8)
summary(pc_verbal_categories_8)
pc_verbal_categories_8$warnings
pc_verbal_categories_8$errors
```

```{r PowerCurve_m2_verbal_cat8_add_to_df}
results <- summary(pc_verbal_categories_8)
results_df <- as.data.frame(cbind(rep("verbal_s30_cat8", 
                                        times= results$trials), 
                                    rep(30, times = results$trials),
                                    rep(37, times = results$trials),
                                    rep(8, times = results$trials),
              pc_verbal_categories_8[["ps"]][[1]]$pval, 
              rep(results$successes, times = results$trials),
              rep(results$mean,  times = results$trials),
              rep(results$lower,  times = results$trials),
              rep(results$upper,  times = results$trials),
              rep(nrow(pc_verbal_categories_8$warnings), 
                  times = results$trials)))
colnames(results_df)<-c("name", "nsubjects", "ntrials", "ncat",
                        "pval","successes", "mean", "lower", "upper", "warnings")
``` 

```{r PowerCurve_m2_verbal_cat8_export}
simr_results <- rbind(simr_results, results_df)
write.csv(simr_results, here::here("results", "power_sim",  "posthoc_powersim.csv"))
```

*30 participants, 6 categories:*

```{r PowerCurve_m2_verbal_cat6,results="hide", cache = TRUE}
(pc_verbal_categories_6 <- powerCurve(m_verbal, test=fixed("Pos.cont", method = "t"),
                                       along = "category", breaks = c(6),
                                       nsim = 1000, seed = 99))

saveRDS(pc_verbal_categories_6, 
        file = here::here("results", "power_sim",
                          "pc_verbal_categories_6.RDS"))
```

```{r PowerCurve_m2_verbal_cat6_summary}
#plot(pc_verbal_categories_6)
summary(pc_verbal_categories_6)
pc_verbal_categories_6$warnings
pc_verbal_categories_6$errors
```

```{r PowerCurve_m2_verbal_cat6_add_to_df}
results <- summary(pc_verbal_categories_6)
results_df <- as.data.frame(cbind(rep("verbal_s30_cat6", 
                                        times= results$trials), 
                                    rep(30, times = results$trials),
                                    rep(27, times = results$trials),
                                    rep(6, times = results$trials),
              pc_verbal_categories_6[["ps"]][[1]]$pval, 
              rep(results$successes, times = results$trials),
              rep(results$mean,  times = results$trials),
              rep(results$lower,  times = results$trials),
              rep(results$upper,  times = results$trials),
              rep(nrow(pc_verbal_categories_6$warnings), 
                  times = results$trials)))
colnames(results_df)<-c("name", "nsubjects", "ntrials", "ncat",
                        "pval","successes", "mean", "lower", "upper", "warnings")
``` 

```{r PowerCurve_m2_verbal_cat6_export}
simr_results <- rbind(simr_results, results_df)
write.csv(simr_results, here::here("results", "power_sim",  "posthoc_powersim.csv"))
```


*30 participants, 4 categories:*

```{r PowerCurve_m2_verbal_cat4,results="hide", cache = TRUE}
(pc_verbal_categories_4 <- powerCurve(m_verbal, test=fixed("Pos.cont", method = "t"),
                                       along = "category", breaks = c(4),
                                       nsim = 1000, seed = 99))

saveRDS(pc_verbal_categories_4, 
        file = here::here("results", "power_sim",
                          "pc_verbal_categories_4.RDS"))
```

```{r PowerCurve_m2_verbal_cat4_summary}
#plot(pc_verbal_categories_4)
summary(pc_verbal_categories_4)
pc_verbal_categories_4$warnings
pc_verbal_categories_4$errors
```

```{r PowerCurve_m2_verbal_cat4_add_to_df}
results <- summary(pc_verbal_categories_4)
results_df <- as.data.frame(cbind(rep("verbal_s30_cat4", 
                                        times= results$trials), 
                                    rep(30, times = results$trials),
                                    rep(19, times = results$trials),
                                    rep(4, times = results$trials),
              pc_verbal_categories_4[["ps"]][[1]]$pval, 
              rep(results$successes, times = results$trials),
              rep(results$mean,  times = results$trials),
              rep(results$lower,  times = results$trials),
              rep(results$upper,  times = results$trials),
              rep(nrow(pc_verbal_categories_4$warnings), 
                  times = results$trials)))
colnames(results_df)<-c("name", "nsubjects", "ntrials", "ncat",
                        "pval","successes", "mean", "lower", "upper", "warnings")
``` 

```{r PowerCurve_m2_verbal_cat4_export}
simr_results <- rbind(simr_results, results_df)
write.csv(simr_results, here::here("results", "power_sim",  "posthoc_powersim.csv"))
```




#### A) **TYPING DATA**

*30 participants, 120 trials - The actual post-hoc power:*  

```{r PowerSim_m2_typing_30,results="hide", cache = TRUE}
(Power_m2_typing_30 <- powerSim(m_typing, 
                                test=fixed("Pos.cont", method = "t"),
                                nsim = 1000)) # increase to nsim = 1000 for real test

saveRDS(Power_m2_typing_30, 
        file = here::here("results", "power_sim",
                          "pc_typing30.RDS"))


results <- summary(Power_m2_typing_30)
lastResult()$warnings
lastResult()$errors

``` 

```{r PowerCurve_m2_typing_subj30_add_to_df}
results_df <- as.data.frame(cbind(rep("typing_s30_t_around109", 
                                        times= results$trials), 
                                    rep(30, times = results$trials),
                                    rep(109, times = results$trials),
                                    rep(24, times = results$trials),
                Power_m2_typing_30[["pval"]], 
              rep(results$successes, times = results$trials),
              rep(results$mean,  times = results$trials),
              rep(results$lower,  times = results$trials),
              rep(results$upper,  times = results$trials),
              rep(nrow(Power_m2_typing_30$warnings), 
                  times = results$trials)))
colnames(results_df)<-c("name", "nsubjects", "ntrials", "ncat",
                        "pval","successes", "mean", "lower", "upper", "warnings")
``` 

```{r PowerCurve_m2_typing_subj30_export}
simr_results <- rbind(simr_results, results_df)
write.csv(simr_results, here::here("results", "power_sim",  "posthoc_powersim.csv"))
```


*24 participants, 120 trials:*


```{r PowerCurve_m2_typing_24,results="hide", cache = TRUE}
(pc_typing24 <- powerCurve(m_typing, test=fixed("Pos.cont", method = "t"),
                           along = "subject", breaks = c(24), nsim = 1000))
saveRDS(pc_typing24, 
        file = here::here("results", "power_sim",
                          "pc_typing24.RDS"))
```

```{r PowerCurve_m2_typing_24_results}
#plot(pc_typing24)
summary(pc_typing24)
pc_typing24$warnings
pc_typing24$errors
```

```{r PowerCurve_m2_typing_24_add_to_df}
results <- summary(pc_typing24)
results_df <- as.data.frame(cbind(rep("typing_s24_t_around109", 
                                        times= results$trials), 
                                    rep(24, times = results$trials),
                                    rep(87, times = results$trials),
                                  rep(24, times = results$trials),
              pc_typing24[["ps"]][[1]]$pval, 
              rep(results$successes, times = results$trials),
              rep(results$mean,  times = results$trials),
              rep(results$lower,  times = results$trials),
              rep(results$upper,  times = results$trials),
              rep(nrow(pc_typing24$warnings), 
                  times = results$trials)))
colnames(results_df)<-c("name", "nsubjects", "ntrials", "ncat",
                        "pval","successes", "mean", "lower", "upper", "warnings")
``` 

```{r PowerCurve_m2_typing_24_export}
simr_results <- rbind(simr_results, results_df)
write.csv(simr_results, here::here("results", "power_sim",  "posthoc_powersim.csv"))
```


*20 participants, 120 trials:*


```{r PowerCurve_m2_typing_20,results="hide", cache = TRUE}
(pc_typing20 <- powerCurve(m_typing, test=fixed("Pos.cont", method = "t"),
                           along = "subject", breaks = c(20), nsim = 1000))
saveRDS(pc_typing20, 
        file = here::here("results", "power_sim",
                          "pc_typing20.RDS"))
```

```{r PowerCurve_m2_typing_20_results}
#plot(pc_typing20)
summary(pc_typing20)
pc_typing20$warnings
pc_typing20$errors
```

```{r PowerCurve_m2_typing_20_add_to_df}
results <- summary(pc_typing20)
results_df <- as.data.frame(cbind(rep("typing_s20_t_around109", 
                                        times= results$trials), 
                                    rep(20, times = results$trials),
                                    rep(73, times = results$trials),
                                  rep(24, times = results$trials),
              pc_typing20[["ps"]][[1]]$pval, 
              rep(results$successes, times = results$trials),
              rep(results$mean,  times = results$trials),
              rep(results$lower,  times = results$trials),
              rep(results$upper,  times = results$trials),
              rep(nrow(pc_typing20$warnings), 
                  times = results$trials)))
colnames(results_df)<-c("name", "nsubjects", "ntrials", "ncat",
                        "pval","successes", "mean", "lower", "upper", "warnings")
``` 

```{r PowerCurve_m2_typing_20_export}
simr_results <- rbind(simr_results, results_df)
write.csv(simr_results, here::here("results", "power_sim",  "posthoc_powersim.csv"))
```


*16 participants, 120 trials:*

```{r PowerCurve_m2_typing_16,results="hide", cache = TRUE}
(pc_typing16 <- powerCurve(m_typing, test=fixed("Pos.cont", method = "t"),
                           along = "subject", breaks = c(16), nsim = 1000))
saveRDS(pc_typing16, 
        file = here::here("results", "power_sim",
                          "pc_typing16.RDS"))
```

```{r PowerCurve_m2_typing_16_summary}
#plot(pc_typing16)
summary(pc_typing16)
pc_typing16$warnings
pc_typing16$errors
```

```{r PowerCurve_m2_typing_16_add_to_df}
results <- summary(pc_typing16)
results_df <- as.data.frame(cbind(rep("typing_s16_t_around109", 
                                        times= results$trials), 
                                    rep(16, times = results$trials),
                                    rep(58, times = results$trials),
                                rep(24, times = results$trials),
              pc_typing16[["ps"]][[1]]$pval, 
              rep(results$successes, times = results$trials),
              rep(results$mean,  times = results$trials),
              rep(results$lower,  times = results$trials),
              rep(results$upper,  times = results$trials),
              rep(nrow(pc_typing16$warnings), 
                  times = results$trials)))
colnames(results_df)<-c("name", "nsubjects", "ntrials", "ncat",
                        "pval","successes", "mean", "lower", "upper", "warnings")
``` 

```{r PowerCurve_m2_typing_16_export}
simr_results <- rbind(simr_results, results_df)
write.csv(simr_results, here::here("results", "power_sim",  "posthoc_powersim.csv"))
```


*10 participants, 120 trials:*


```{r PowerCurve_m2_typing_10,results="hide", cache = TRUE}
(pc_typing10 <- powerCurve(m_typing, test=fixed("Pos.cont", method = "t"),
                           along = "subject", breaks = c(10), nsim = 1000))
saveRDS(pc_typing10, 
        file = here::here("results", "power_sim",
                          "pc_typing10.RDS"))
```

```{r PowerCurve_m2_typing_10_summary}
#plot(pc_typing10)
summary(pc_typing10)
pc_typing10$warnings
pc_typing10$errors
```

```{r PowerCurve_m2_typing_10_add_to_df}
results <- summary(pc_typing10)
results_df <- as.data.frame(cbind(rep("typing_s10_t_around109", 
                                        times= results$trials), 
                                    rep(10, times = results$trials),
                                    rep(36, times = results$trials),
                                  rep(24, times = results$trials),
              pc_typing10[["ps"]][[1]]$pval, 
              rep(results$successes, times = results$trials),
              rep(results$mean,  times = results$trials),
              rep(results$lower,  times = results$trials),
              rep(results$upper,  times = results$trials),
              rep(nrow(pc_typing10$warnings), 
                  times = results$trials)))
colnames(results_df)<-c("name", "nsubjects", "ntrials", "ncat",
                        "pval","successes", "mean", "lower", "upper", "warnings")
``` 

```{r PowerCurve_m2_typing_10_export}
simr_results <- rbind(simr_results, results_df)
write.csv(simr_results, here::here("results", "power_sim",  "posthoc_powersim.csv"))
```



*8 participants, 120 trials:*

```{r PowerCurve_m2_typing_8,results="hide", cache = TRUE}
(pc_typing8 <- powerCurve(m_typing, test=fixed("Pos.cont", method = "t"),
                           along = "subject", breaks = c(8), nsim = 1000))
saveRDS(pc_typing8, 
        file = here::here("results", "power_sim",
                          "pc_typing8.RDS"))
```

```{r PowerCurve_m2_typing_8_summary}
#plot(pc_typing8)
summary(pc_typing8)
pc_typing8$warnings
pc_typing8$errors
```

```{r PowerCurve_m2_typing_8_add_to_df}
results <- summary(pc_typing8)
results_df <- as.data.frame(cbind(rep("typing_s8_t_around109", 
                                        times= results$trials), 
                                    rep(8, times = results$trials),
                                    rep(109, times = results$trials),
                                  rep(24, times = results$trials),
              pc_typing8[["ps"]][[1]]$pval, 
              rep(results$successes, times = results$trials),
              rep(results$mean,  times = results$trials),
              rep(results$lower,  times = results$trials),
              rep(results$upper,  times = results$trials),
              rep(nrow(pc_typing8$warnings), 
                  times = results$trials)))
colnames(results_df)<-c("name", "nsubjects", "ntrials", "ncat", 
                        "pval","successes", "mean", "lower", "upper", "warnings")
``` 

```{r PowerCurve_m2_typing_8_export}
simr_results <- rbind(simr_results, results_df)
write.csv(simr_results, here::here("results", "power_sim",  "posthoc_powersim.csv"))
```


*6 participants, 120 trials:*


```{r PowerCurve_m2_typing_6,results="hide", cache = TRUE}
(pc_typing6 <- powerCurve(m_typing, test=fixed("Pos.cont", method = "t"),
                           along = "subject", breaks = c(6), nsim = 1000))
saveRDS(pc_typing6, 
        file = here::here("results", "power_sim",
                          "pc_typing6.RDS"))
```

```{r PowerCurve_m2_typing_6_summary}
#plot(pc_typing6)
summary(pc_typing6)
pc_typing6$warnings
pc_typing6$errors
```

```{r PowerCurve_m2_typing_6_add_to_df}
results <- summary(pc_typing6)
results_df <- as.data.frame(cbind(rep("typing_s6_t_around109", 
                                        times= results$trials), 
                                    rep(6, times = results$trials),
                                    rep(18, times = results$trials),
                                  rep(24, times = results$trials),
              pc_typing6[["ps"]][[1]]$pval, 
              rep(results$successes, times = results$trials),
              rep(results$mean,  times = results$trials),
              rep(results$lower,  times = results$trials),
              rep(results$upper,  times = results$trials),
              rep(nrow(pc_typing6$warnings), 
                  times = results$trials)))
colnames(results_df)<-c("name", "nsubjects", "ntrials", "ncat",
                        "pval","successes", "mean", "lower", "upper", "warnings")
``` 

```{r PowerCurve_m2_typing_6_export}
simr_results <- rbind(simr_results, results_df)
write.csv(simr_results, here::here("results", "power_sim",  "posthoc_powersim.csv"))
```


*30 participants, 19 categories:*

```{r PowerCurve_m2_typing_cat19,results="hide", cache = TRUE}
(pc_typing_categories_19 <- powerCurve(m_typing, test=fixed("Pos.cont", method = "t"),
                                       along = "category", breaks = c(19),
                                       nsim = 1000, seed = 99))
saveRDS(pc_typing_categories_19, 
        file = here::here("results", "power_sim",
                          "pc_typing_categories_19.RDS"))
```

```{r PowerCurve_m2_typing_cat19_summary}
#plot(pc_typing_categories_19)
summary(pc_typing_categories_19)
pc_typing_categories_19$warnings
pc_typing_categories_19$errors
```

```{r PowerCurve_m2_typing_cat19_add_to_df}
results <- summary(pc_typing_categories_19)
results_df <- as.data.frame(cbind(rep("typing_s30_cat19", 
                                        times= results$trials), 
                                    rep(30, times = results$trials),
                                    rep(86, times = results$trials),
                                    rep(19, times = results$trials),
              pc_typing_categories_19[["ps"]][[1]]$pval, 
              rep(results$successes, times = results$trials),
              rep(results$mean,  times = results$trials),
              rep(results$lower,  times = results$trials),
              rep(results$upper,  times = results$trials),
              rep(nrow(pc_typing_categories_19$warnings), 
                  times = results$trials)))
colnames(results_df)<-c("name", "nsubjects", "ntrials", "ncat",
                        "pval","successes", "mean", "lower", "upper", "warnings")
``` 

```{r PowerCurve_m2_typing_cat19_export}
simr_results <- rbind(simr_results, results_df)
write.csv(simr_results, here::here("results", "power_sim",  "posthoc_powersim.csv"))
```


*30 participants, 16 categories:*

```{r PowerCurve_m2_typing_cat16,results="hide", cache = TRUE}
(pc_typing_categories_16 <- powerCurve(m_typing, test=fixed("Pos.cont", method = "t"),
                                       along = "category", breaks = c(16),
                                       nsim = 1000, seed = 99))
saveRDS(pc_typing_categories_16, 
        file = here::here("results", "power_sim",
                          "pc_typing_categories_16.RDS"))
```

```{r PowerCurve_m2_typing_cat16_summary}
#plot(pc_typing_categories_16)
summary(pc_typing_categories_16)
pc_typing_categories_16$warnings
pc_typing_categories_16$errors
```

```{r PowerCurve_m2_typing_cat16_add_to_df}
results <- summary(pc_typing_categories_16)
results_df <- as.data.frame(cbind(rep("typing_s30_cat16", 
                                        times= results$trials), 
                                    rep(30, times = results$trials),
                                    rep(73, times = results$trials),
                                    rep(16, times = results$trials),
              pc_typing_categories_16[["ps"]][[1]]$pval, 
              rep(results$successes, times = results$trials),
              rep(results$mean,  times = results$trials),
              rep(results$lower,  times = results$trials),
              rep(results$upper,  times = results$trials),
              rep(nrow(pc_typing_categories_16$warnings), 
                  times = results$trials)))
colnames(results_df)<-c("name", "nsubjects", "ntrials", "ncat",
                        "pval","successes", "mean", "lower", "upper", "warnings")
``` 

```{r PowerCurve_m2_typing_cat16_export}
simr_results <- rbind(simr_results, results_df)
write.csv(simr_results, here::here("results", "power_sim",  "posthoc_powersim.csv"))
```


*30 participants, 13 categories:*

```{r PowerCurve_m2_typing_cat13,results="hide", cache = TRUE}
(pc_typing_categories_13 <- powerCurve(m_typing, test=fixed("Pos.cont", method = "t"),
                                       along = "category", breaks = c(13),
                                       nsim = 1000, seed = 99))
saveRDS(pc_typing_categories_13, 
        file = here::here("results", "power_sim",
                          "pc_typing_categories_13.RDS"))
```

```{r PowerCurve_m2_typing_cat13_summary}
#plot(pc_typing_categories_13)
summary(pc_typing_categories_13)
pc_typing_categories_13$warnings
pc_typing_categories_13$errors
```

```{r PowerCurve_m2_typing_cat13_add_to_df}
results <- summary(pc_typing_categories_13)
results_df <- as.data.frame(cbind(rep("typing_s30_cat13", 
                                        times= results$trials), 
                                    rep(30, times = results$trials),
                                    rep(59, times = results$trials),
                                    rep(13, times = results$trials),
              pc_typing_categories_13[["ps"]][[1]]$pval, 
              rep(results$successes, times = results$trials),
              rep(results$mean,  times = results$trials),
              rep(results$lower,  times = results$trials),
              rep(results$upper,  times = results$trials),
              rep(nrow(pc_typing_categories_13$warnings), 
                  times = results$trials)))
colnames(results_df)<-c("name", "nsubjects", "ntrials", "ncat",
                        "pval","successes", "mean", "lower", "upper", "warnings")
``` 

```{r PowerCurve_m2_typing_cat13_export}
simr_results <- rbind(simr_results, results_df)
write.csv(simr_results, here::here("results", "power_sim",  "posthoc_powersim.csv"))
```


*30 participants, 8 categories:*

```{r PowerCurve_m2_typing_cat8,results="hide", cache = TRUE}
(pc_typing_categories_8 <- powerCurve(m_typing, test=fixed("Pos.cont", method = "t"),
                                       along = "category", breaks = c(8),
                                       nsim = 1000, seed = 99))
saveRDS(pc_typing_categories_8, 
        file = here::here("results", "power_sim",
                          "pc_typing_categories_8.RDS"))
```

```{r PowerCurve_m2_typing_cat8_summary}
#plot(pc_typing_categories_8)
summary(pc_typing_categories_8)
pc_typing_categories_8$warnings
pc_typing_categories_8$errors
```

```{r PowerCurve_m2_typing_cat8_add_to_df}
results <- summary(pc_typing_categories_8)
results_df <- as.data.frame(cbind(rep("typing_s30_cat8", 
                                        times= results$trials), 
                                    rep(30, times = results$trials),
                                    rep(37, times = results$trials),
                                    rep(8, times = results$trials),
              pc_typing_categories_8[["ps"]][[1]]$pval, 
              rep(results$successes, times = results$trials),
              rep(results$mean,  times = results$trials),
              rep(results$lower,  times = results$trials),
              rep(results$upper,  times = results$trials),
              rep(nrow(pc_typing_categories_8$warnings), 
                  times = results$trials)))
colnames(results_df)<-c("name", "nsubjects", "ntrials", "ncat",
                        "pval","successes", "mean", "lower", "upper", "warnings")
``` 

```{r PowerCurve_m2_typing_cat8_export}
simr_results <- rbind(simr_results, results_df)
write.csv(simr_results, here::here("results", "power_sim",  "posthoc_powersim.csv"))
```


*30 participants, 6 categories:*

```{r PowerCurve_m2_typing_cat6,results="hide", cache = TRUE}
(pc_typing_categories_6 <- powerCurve(m_typing, test=fixed("Pos.cont", method = "t"),
                                       along = "category", breaks = c(6),
                                       nsim = 1000, seed = 99))
saveRDS(pc_typing_categories_6, 
        file = here::here("results", "power_sim",
                          "pc_typing_categories_6.RDS"))
```

```{r PowerCurve_m2_typing_cat6_summary}
#plot(pc_typing_categories_6)
summary(pc_typing_categories_6)
pc_typing_categories_6$warnings
pc_typing_categories_6$errors
```

```{r PowerCurve_m2_typing_cat6_add_to_df}
results <- summary(pc_typing_categories_6)
results_df <- as.data.frame(cbind(rep("typing_s30_cat6", 
                                        times= results$trials), 
                                    rep(30, times = results$trials),
                                    rep(73, times = results$trials),
                                    rep(27, times = results$trials),
              pc_typing_categories_6[["ps"]][[1]]$pval, 
              rep(results$successes, times = results$trials),
              rep(results$mean,  times = results$trials),
              rep(results$lower,  times = results$trials),
              rep(results$upper,  times = results$trials),
              rep(nrow(pc_typing_categories_6$warnings), 
                  times = results$trials)))
colnames(results_df)<-c("name", "nsubjects", "ntrials", "ncat",
                        "pval","successes", "mean", "lower", "upper", "warnings")
``` 

```{r PowerCurve_m2_typing_cat6_export}
simr_results <- rbind(simr_results, results_df)
write.csv(simr_results, here::here("results", "power_sim",  "posthoc_powersim.csv"))
```

*30 participants, 4 categories:*

```{r PowerCurve_m2_typing_cat4,results="hide", cache = TRUE}
(pc_typing_categories_4 <- powerCurve(m_typing, test=fixed("Pos.cont", method = "t"),
                                       along = "category", breaks = c(4),
                                       nsim = 1000, seed = 99))
saveRDS(pc_typing_categories_4, 
        file = here::here("results", "power_sim",
                          "pc_typing_categories_4.RDS"))
```

```{r PowerCurve_m2_typing_cat4_summary}
#plot(pc_typing_categories_4)
summary(pc_typing_categories_4)
pc_typing_categories_4$warnings
pc_typing_categories_4$errors
```

```{r PowerCurve_m2_typing_cat4_add_to_df}
results <- summary(pc_typing_categories_4)
results_df <- as.data.frame(cbind(rep("typing_s30_cat4", 
                                        times= results$trials), 
                                    rep(30, times = results$trials),
                                    rep(19, times = results$trials),
                                    rep(4, times = results$trials),
              pc_typing_categories_4[["ps"]][[1]]$pval, 
              rep(results$successes, times = results$trials),
              rep(results$mean,  times = results$trials),
              rep(results$lower,  times = results$trials),
              rep(results$upper,  times = results$trials),
               rep(nrow(pc_typing_categories_4$warnings), 
                  times = results$trials)))
colnames(results_df)<-c("name", "nsubjects", "ntrials", "ncat",
                        "pval","successes", "mean", "lower", "upper", "warnings")
``` 

```{r PowerCurve_m2_typing_cat4_export}
simr_results <- rbind(simr_results, results_df)
write.csv(simr_results, here::here("results", "power_sim",  "posthoc_powersim.csv"))
```
